requirements.txt
pandas
streamlit
numpy
openai
python-dotenv
